backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
pymongo>=4.0
python-dotenv>=0.20
reportlab>=3.6
pdfplumber>=0.7
